move `getGitHubUser` into standalone module
