+++ lib/github-package.js
@@ -24,9 +24,7 @@ import WorkerManager from './worker-manager';
 import getRepoPipelineManager from './get-repo-pipeline-manager';
 import {reporterProxy} from './reporter-proxy';
 import ModelObserver from './models/model-observer';
-// FIXME: move the following dependencies to new file when you extract the graphQL stuff
-import RelayNetworkLayerManager from './relay-network-layer-manager';
-import {UNAUTHENTICATED, INSUFFICIENT} from './shared/keytar-strategy';
+import getGitHubUser from './github-user';
 
 const defaultState = {
   newProject: true,
@@ -105,38 +103,13 @@ export default class GithubPackage {
 
     this.loginObserver = new ModelObserver({
       didUpdate: async () => {
-        const gitHubUser = await this.getGitHubUser();
+        const gitHubUser = await getGitHubUser(this.loginModel);
         reporterProxy.setGitHubUser(gitHubUser);
       }
     });
     this.loginObserver.setActiveModel(this.loginModel);
 
     this.setupYardstick();
-  }
-
-  async getGitHubUser() {
-    const token = await this.loginModel.getToken('https://api.github.com');
-    if (token === UNAUTHENTICATED || token === INSUFFICIENT) {
-      return null;
-    }
-
-    const fetchQuery = RelayNetworkLayerManager.getFetchQuery('https://api.github.com/graphql', token);
-    const response = await fetchQuery({
-      name: 'GetGitHubUser',
-      text: `
-      query {
-        viewer {
-          login
-            }
-          }
-    `,
-    });
-    if (response.errors && response.errors.length > 1) {
-      // eslint-disable-next-line no-console
-      console.error(`Error fetching GitHub username:\n${response.errors.map(e => e.message).join('\n')}`);
-    }
-    // do I need to guard against data not containing "viewer" here?
-    return response.data.viewer.login;
   }
 
   setupYardstick() {
